Implementing scripts for stationary states (MF and MC) + implementation of complex contagion MF

+++ dynalearn/meanfields/simple_contagion.py
@@ -4,10 +4,12 @@ from .mf import *
 
 
 class SIS_AME(AME):
-    def __init__(self, p_k, infection_prob, recovery_prob, tol=1e-3, verbose=1):
+    def __init__(
+        self, p_k, infection_prob, recovery_prob, tol=1e-3, verbose=1, dtype="float"
+    ):
         self.infection_prob = infection_prob
         self.recovery_prob = recovery_prob
-        super(SIS_AME, self).__init__(2, p_k, tol=tol, verbose=verbose)
+        super(SIS_AME, self).__init__(2, p_k, tol=tol, verbose=verbose, dtype=dtype)
 
     def compute_ltp(self,):
         ltp = np.zeros((self.s_dim, self.s_dim, *self.k_grid.shape))
@@ -19,10 +21,12 @@ class SIS_AME(AME):
 
 
 class SIS_PA(PA):
-    def __init__(self, p_k, infection_prob, recovery_prob, tol=1e-3, verbose=1):
+    def __init__(
+        self, p_k, infection_prob, recovery_prob, tol=1e-3, verbose=1, dtype="float"
+    ):
         self.infection_prob = infection_prob
         self.recovery_prob = recovery_prob
-        super(SIS_PA, self).__init__(2, p_k, tol=tol, verbose=verbose)
+        super(SIS_PA, self).__init__(2, p_k, tol=tol, verbose=verbose, dtype=dtype)
 
     def compute_ltp(self,):
         ltp = np.zeros((self.s_dim, self.s_dim, *self.k_grid.shape))
@@ -34,10 +38,12 @@ class SIS_PA(PA):
 
 
 class SIS_MF(MF):
-    def __init__(self, p_k, infection_prob, recovery_prob, tol=1e-3, verbose=1):
+    def __init__(
+        self, p_k, infection_prob, recovery_prob, tol=1e-3, verbose=1, dtype="float"
+    ):
         self.infection_prob = infection_prob
         self.recovery_prob = recovery_prob
-        super(SIS_MF, self).__init__(2, p_k, tol=tol, verbose=verbose)
+        super(SIS_MF, self).__init__(2, p_k, tol=tol, verbose=verbose, dtype=dtype)
 
     def compute_ltp(self,):
         ltp = np.zeros((self.s_dim, self.s_dim, *self.k_grid.shape))
@@ -49,10 +55,12 @@ class SIS_MF(MF):
 
 
 class SIR_AME(AME):
-    def __init__(self, p_k, infection_prob, recovery_prob, tol=1e-3, verbose=1):
+    def __init__(
+        self, p_k, infection_prob, recovery_prob, tol=1e-3, verbose=1, dtype="float"
+    ):
         self.infection_prob = infection_prob
         self.recovery_prob = recovery_prob
-        super(SIR_AME, self).__init__(3, p_k, tol=tol, verbose=verbose)
+        super(SIR_AME, self).__init__(3, p_k, tol=tol, verbose=verbose, dtype=dtype)
 
     def compute_ltp(self,):
         ltp = np.zeros((self.s_dim, self.s_dim, *self.k_grid.shape))
@@ -69,10 +77,12 @@ class SIR_AME(AME):
 
 
 class SIR_PA(PA):
-    def __init__(self, p_k, infection_prob, recovery_prob, tol=1e-3, verbose=1):
+    def __init__(
+        self, p_k, infection_prob, recovery_prob, tol=1e-3, verbose=1, dtype="float"
+    ):
         self.infection_prob = infection_prob
         self.recovery_prob = recovery_prob
-        super(SIR_PA, self).__init__(3, p_k, tol=tol, verbose=verbose)
+        super(SIR_PA, self).__init__(3, p_k, tol=tol, verbose=verbose, dtype=dtype)
 
     def compute_ltp(self,):
         ltp = np.zeros((self.s_dim, self.s_dim, *self.k_grid.shape))
@@ -89,10 +99,12 @@ class SIR_PA(PA):
 
 
 class SIR_MF(MF):
-    def __init__(self, p_k, infection_prob, recovery_prob, tol=1e-3, verbose=1):
+    def __init__(
+        self, p_k, infection_prob, recovery_prob, tol=1e-3, verbose=1, dtype="float"
+    ):
         self.infection_prob = infection_prob
         self.recovery_prob = recovery_prob
-        super(SIR_MF, self).__init__(3, p_k, tol=tol, verbose=verbose)
+        super(SIR_MF, self).__init__(3, p_k, tol=tol, verbose=verbose, dtype=dtype)
 
     def compute_ltp(self,):
         tp = np.zeros((self.s_dim, self.s_dim, *self.k_grid.shape))
